[TestLoop Refactor] (1/n) Overhaul the sender framework. (#10614)
Previously, we had CanSend and CanSendAsync traits, and corresponding to
them, the Sender and AsyncSender wrapper structs. The difference between
the two types of senders turns out to be a little tricky to work with,
so now there is only one Sender, whereas `AsyncSender<M, R>` is now an
alias for `Sender<MessageExpectingResponse<M, R>>`.

Another pain point was the ability to combine several senders into one;
PeerManagerAdapter is an example; previously we had to use some
boilerplate to make it convertible from any object that implements all
of its CanSend and CanSendAsync requirements, and then to do such
conversion, we used `.into()`, which is quite confusing -- somehow, an
`Addr<PeerManagerActor>` can be `.into()`-ed into a
`PeerManagerAdapter`. Now, instead of all that, we use proc macros.
`PeerManagerAdapter` is annotated with `#[derive(MultiSend,
MultiSenderFrom)]`, the former allows PeerManagerAdapter to be used like
any of the senders it contains, and the latter allows
`Addr<PeerManagerActor>` to be converted to it via
`.into_multi_sender()`. This interface is more explicit and doesn't use
the confusing `.into()`.

There are also some general QoL improvements. The interface on
`LateBoundSender` and `noop()` is also improved. Wrapping things into
`Arc`s is generally no longer needed now.
`Sender::from_fn`/`Sender::from_async_fn` can be used to create a sender
(for testing) from an inline function.

The rest of the PR is just adapting the existing usages of
Sender/AsyncSender to the new API.