fix: don't panic when a Client receives ChunkStateWitness with invalid shard_id (#10621)
When Client start processing a new `ChunkStateWitness` which it
received, it immediately fetches the previous block and the previous
chunk using `prev_block_hash` and `shard_id` provided by the
`ChunkStateWitness`.
It turns out that the current implementation of
`Chain::get_prev_chunk_header` will panic if it's given an invalid
`shard_id`, which means that a malicious peer could send a
`ChunkStateWitness` with a bad `shard_id` and it would crash the node
that received it.


https://github.com/near/nearcore/blob/c5c84adbadf54957083a33ec243cb334b411daee/chain/client/src/stateless_validation/chunk_validator.rs#L596-L607

`Chain::get_prev_chunk_header` is very deceptive, it returns a `Result`,
but still does an `unwrap()` inside. Let's fix the problem by removing
the `unwrap()`. From now on `Chain::get_prev_chunk_header` will return a
error when it encounters an invalid `shard_id`.

A test is added to ensure that this bug doesn't happen again.