[TestLoop refactoring] (2/n) Use Sender framework for the network stack and jsonrpc stack (#10627)
Quite a large change. Here's the idea:

* Introduce `ClientSenderForNetwork` and `ClientSenderForRpc`, which are
multisenders. `ClientSenderForNetwork` is used by the network layer to
send messages to the client, and `ClientSenderForRpc` is used by the
jsonrpc layer to send messages to the client. These layers no longer
require a `Addr<ClientActor>` or `Addr<ViewClientActor>` to construct.
* The rest of the changes are just whatever that is necessary to make
this work:
* There is no longer a trait abstraction
(`near_network::client::Client`) for the network to call the client.
This trait abstraction is nice, but it introduces an additional
indirection to make things confusing, and is also arbitrarily designed
(there's non-uniform data transformation for the requests and
responses), and not consistent with the rest of the codebase. Instead,
usages of this trait are replaced with direct calls to the
`ClientSenderForNetwork` multisender.
* That means that the struct being sent from the network layer to the
client is now a part of the `ClientSenderForNetwork` interface, meaning
that I have moved these structs into the near_network crate. This is not
ideal because these structs are annotated with actix response types,
making them somewhat "actix dependent". It could be avoided by
introducing intermediate structs but that is a bit of an overkill.
Instead, I just make each sender in `ClientSenderForNetwork` an
`AsyncSender<M, ActixResult<M>>`; it's convenient (coz we don't need to
specify the return result) and guaranteed to work with the actix side.
In the future if we are to get rid of actix, we just replace this with
the spelling of the real type.
* Another implementation of the old `near_network::client::Client` is
`FakeClient`. This is quite annoying because for two messages it
responds with a sane response, but for other messages it wraps them and
sends them into a `Sink`. Now, `Sink` is another abstraction that
basically has the same semantics as `Sender` so I've replaced all usages
of it with `Sender`, removed the entire `FakeClient`, and moved the
custom behavior of handling those two messages into
`peer_manager/testonly.rs`. Also, the events that were sent to the sink
do not appear to be used, I'm not sure what's going with that, perhaps
it was a refactoring that was half done before.
* On the jsonrpc side, all the errors had to be convertible to some kind
of RpcError, and it used to be that we convert actix's `MailboxError` to
each type of RpcError, but now that we're not relying on actix, we
instead convert `AsyncSendError`. The semantics are the same in the end.
* Finally, there's an event_sink in the `config.rs`'s `NetworkConfig`
struct. This was only used for testing to collect various events, but
these events are written to the sink in various places in the networking
code and some of them are not efficient because they require additional
cloning. Previously, the way it was handled was that the Sink had an
extra feature of sending something "lazily" (by calling it with a
closure) so that if the sink was an noop-sink, the closure is not called
at all, thereby avoiding expensive computation in the production code
path. Now, `Sender` does not have such functionality (and adding such
functionality would add extra complexity) and it's not even clear how to
define lazy sending, so instead, I've guarded the whole `event_sink` and
all usages with `#[cfg(test)]`. Again, these do not seem to be used
today.