[TestLoop Refactoring] (3/n) Split ClientActions from ClientActor. (#10646)
This makes it so that ClientActor has no logic, and is purely an Actix
shell. This way, ClientActions can be independently tested.

Also, this adds another example to the core-async library that shows how
to test an Actix component. It demonstrates the use of TestLoop to test
both asynchronous actix requests and periodic timers. The same idea can
be used to test the client actor by constructing and using ClientActions
in a TestLoop test.