[resharding] Handle case of restarting node during resharding catchup (#10611)
This PR is a fix for the issue where when we restart a node doing
resharding and in the catchup phase.

High level issue: When we restart a node in the epoch when resharding in
happening, what happens is we go through the whole state_sync process
from the beginning which includes resharding. Once building of the child
trie is completed, we then do a catchup, apply the split state to the
child tried and delete the split state changes for resharding
[here](https://github.com/near/nearcore/blob/e00dcaa72cfed35831b1e72760d21bb8152f1049/chain/chain/src/chain_update.rs#L399).

Zulip thread link:
https://near.zulipchat.com/#narrow/stream/308695-pagoda.2Fprivate/topic/Problems.20after.20resharding.20restart/near/421312417

This is the implementation of Option 1 in the thread.

The key idea here is to not delete the
`DBCol::StateChangesForSplitStates` during catchup of individual blocks
but rather at the end of the catchup phase. This implies, in case we
restart the node in the middle the catchup, we would still have all the
split state information in `DBCol::StateChangesForSplitStates` and it
wouldn't have been deleted.

TODO: Testing on mocknet.