refactor: wrote log_assert macro inline in primitives and remove o11y dependency (#10666)
This is part of the ongoing effort towards reducing dependencies in
near-primitives.
See https://github.com/near/nearcore/issues/8888 for context.

The new crate near-log-utils is a light weight crate which has the
log_assert and log_assert_fail macros that are used in primitives
helping in dropping the o11y dependency from primitives completely.