refactor: move NightshadeRuntime to chain (#10675)
This PR is a part of getting rid of `KeyValueRuntime` (#10678) and
`MockEpochManager` (#10634).

Currently `NightshadeRuntime` is part of `nearcore` which makes it
impossible to use it in `chain` since `chain` depends on `nearcore`. We
need to use `NightshadeRuntime` as a replacement for `KeyValueRuntime`
in `chain` tests. Ideally it would be great to move it to `node-runtime`
crate, but currently it has a lot of dependencies from `near_chain`
which makes it very challenging, so as the first step let's move it to
`chain`.

This PR includes:
* move `nearcore::runtime` to `near_chain::runtime`
* introduce `NightshadeRuntimeExt` to expose
`NightshadeRuntime::from_config` separately since `NearConfig` is part
of `nearcore`
* move `GenesisExt` test methods directly to `Genesis`, also duplicate
some constants for now (will be refactored separately to reduce size of
this PR).
* move the necessary migrations code to `chain`