refactor: deduplicate config-related code (#10681)
In #10675 some of the constants and testing-related methods from
`nearcore::config` were copied to `near_chain_configs::test_utils`. This
PR removes it from `nearcore::configs` and replaces all the usages to
take it from `near_chain_configs`. Also some of the constants are used
not only for testing, so those are moved directly to
`near_chain_configs`.

Also wrap tests inside `nearcore::config` into `mod tests` because that
is how it should be.

This change doesn't introduce any functional changes to the codebase, it
only moves code around.