refactor: Rename receipt id utils for clarity (#10682)
This PR renames the utility function `create_data_id` to
`create_receipt_id_from_action_hash`.

The receipt id is a concept which exists at the Receipt level and is not
specific to a kind of Receipt (whether action or data). Each id needs to
be globally unique across receipts of all kinds, and the ids used for
action and data receipts share the same properties.

There is no meaningful distinction to be made between "data ids" and
other kinds of receipt ids, so I find the name of this low-level utility
to be misleading. In particular, it would not be incorrect to generate
an id for an action receipt using the function in question.

Renaming the function better reflects what it actually does rather than
how it happens to be used.