yield/resume: simplify trie state (#10689)
This PR moves PromiseYield receipts to their own DB column, whereas
previously they were commingled into the postponed receipts column.

Trie state for postponed receipts is complex because it supports a
many-to-one relationship between data receipts and action receipts. It
involves:
* `POSTPONED_RECEIPT_ID`: Mapping from data id to the id of the action
receipt which consumes it.
* `PENDING_DATA_COUNT`: Mapping from action receipt id to the number of
missing data inputs for it.
* `POSTPONED_RECEIPT`: Mapping from action receipt id to the Action
receipt itself.

Because yielded promise always have exactly one input data dependency,
it is sufficient to maintain:
* `PROMISE_YIELD_RECEIPT`: Mapping from data id to the PromiseYield
receipt which consumes it.

Along with the state re-organization, the logic for processing
PromiseYield and PromiseResume receipts is made explicit and separate
from the logic for Action and Data receipts.

---

The changes in this PR bring a few benefits:
- The number of trie keys created and removed over the lifetime of a
yielded promise is reduced.
- The protocol is now robust to delivery of redundant PromiseResume
receipts in the case that a user makes multiple calls to
`promise_yield_resume` with the same data id.
- We are able to eliminate the `YIELED_PROMISE` DB column, which was
previous needed to enforce "at most once" delivery of YieldResume
receipts.

---

Note that the DB layout changes are BC-breaking, but it's OK because
yield execution is still a nightly feature.

I think this PR may break the case in which `promise_yield_create` and
`promise_yield_resume` are called within the same transaction due to the
assumption that the PromiseYield is processed before the PromiseResume.
I plan to completely rework how we handle that case but it turned out to
be more significant than expected. To keep this PR manageable, I'll send
another one with those changes.