fix(resharding): fixed the resharding_rpc_tx nayduck test (#10700)
Something's changed and now when submitting the transaction the result
has status "EXECUTED" and when getting the status later via RPC the
result has status "FINAL". Previously the test was asserting equality of
those, now adjusted to accomodate the different statuses.

First commit is just a minor refactor and the second one is the actual
fix. Best reviewed separately.