Fix incorrect usage of tracing span in async functions. (#10663)
See https://docs.rs/tracing/latest/tracing/struct.Span.html#method.enter

It is incorrect to do this:

```
async fn f() {
  let _ = tracing::span!(...).entered();
  ...
}
```

because when the function yields, the tracing span remains entered. This
causes incorrect span nesting, and we have observed a stack overflow as
a result, when the system is under load.